requirements.txt
langchain>=0.1.0
qdrant-client>=1.1.1
python-docx>=0.8.11
tiktoken>=0.5.0
python-dotenv>=0.19.0
rich>=10.12.0
sentence-transformers>=2.2.2
markdown>=3.4.0
beautifulsoup4>=4.12.0
jieba>=0.42.1 